requirements.txt
dotenv
requests
python-telegram-bot
python-dateutil